move pin poisoning to strightpin class
